Add files via upload
The data given to us is a set of weights (in kilograms) of newborn babies. I loaded this data from a CSV file named 'data7.csv' using NumPy's loadtxt() function and stored it in an array named df_new_born. I then analyzed the distribution of the weights and calculated the mean weight, the 75th percentile weight, and the standard deviation using NumPy.
To describe the distribution of the data, I created a histogram of the newborn weights using Matplotlib's hist() function. The resulting distribution appears to be slightly skewed to the right, indicating that there may be some heavier newborns in the sample.